refactor(index): bundle sracha-index into `sracha index`, gate behind features
Two changes the user pushed for:

1. Drop the standalone `sracha-index` binary. Move its commands
   (extract / build / append / query) into `sracha index <sub>`
   alongside the existing consumer commands (update / status /
   path / clear). Single user-facing binary, single help surface.
   The shared build logic moves from `bin/sracha-index.rs` into
   `sracha_index::build`, hiding the Vortex session plumbing.

2. Make both feature flags off by default until the hosted catalog
   is published:
   - `catalog`: pulls in sracha-index + adds `sracha get
     --catalog DIR` flag (consumer side: lookup + priority seeding)
   - `index`: pulls in `catalog` + exposes the `sracha index`
     subcommand (producer + cache management)

   Default release strips both, dropping ~50 transitive deps
   (vortex stack). Build with `--features index` for the full
   producer toolkit, or `--features catalog` for just the consumer
   `--catalog` flag.

Validated three feature configurations:
- default        — no `--catalog` flag, no `sracha index`
- --features catalog — `--catalog` flag works, no `sracha index`
- --features index   — both work; `sracha index build` + `sracha get
                       --catalog --stream` round-trip produces
                       c1d45b6... (matches disk path)
