+++ crates/sracha-index/Cargo.toml
@@ -27,7 +27,3 @@ sha2 = "0.11"
 vortex = { version = "0.68", features = ["tokio", "unstable_encodings"] }
 vortex-btrblocks = { version = "0.68", features = ["unstable_encodings"] }
 futures = "0.3"
-
-[[bin]]
-name = "sracha-index"
-path = "src/bin/sracha-index.rs"
